Move analysis code into analyser package
